validator works great
